+++ obj/src/auth/BasicAuthorizer.d.ts
@@ -1,4 +1,4 @@
-export declare class BasicAuthManager {
+export declare class BasicAuthorizer {
     anybody(): (req: any, res: any, next: () => void) => void;
     signed(): (req: any, res: any, next: () => void) => void;
 }