+++ obj/src/auth/index.js
@@ -3,12 +3,11 @@ Object.defineProperty(exports, "__esModule", { value: true });
 /**
  * @module auth
  * @preferred
- *
  */
-var BasicAuthManager_1 = require("./BasicAuthManager");
-exports.BasicAuthManager = BasicAuthManager_1.BasicAuthManager;
-var OwnerAuthManager_1 = require("./OwnerAuthManager");
-exports.OwnerAuthManager = OwnerAuthManager_1.OwnerAuthManager;
-var RoleAuthManager_1 = require("./RoleAuthManager");
-exports.RoleAuthManager = RoleAuthManager_1.RoleAuthManager;
+var BasicAuthorizer_1 = require("./BasicAuthorizer");
+exports.BasicAuthManager = BasicAuthorizer_1.BasicAuthorizer;
+var OwnerAuthorizer_1 = require("./OwnerAuthorizer");
+exports.OwnerAuthManager = OwnerAuthorizer_1.OwnerAuthorizer;
+var RoleAuthorizer_1 = require("./RoleAuthorizer");
+exports.RoleAuthManager = RoleAuthorizer_1.RoleAuthorizer;
 //# sourceMappingURL=index.js.map