+++ src/auth/BasicAuthorizer.ts
@@ -4,7 +4,7 @@ const _ = require('lodash');
 import { UnauthorizedException } from 'pip-services3-commons-node';
 import { HttpResponseSender } from '../services/HttpResponseSender';
 
-export class BasicAuthManager {
+export class BasicAuthorizer {
 
     public anybody(): (req: any, res: any, next: () => void) => void {
         return (req, res, next) => {