+++ src/auth/OwnerAuthorizer.ts
@@ -4,7 +4,7 @@ const _ = require('lodash');
 import { UnauthorizedException } from 'pip-services3-commons-node';
 import { HttpResponseSender } from '../services/HttpResponseSender';
 
-export class OwnerAuthManager {
+export class OwnerAuthorizer {
 
     public owner(idParam: string = 'user_id'): (req: any, res: any, next: () => void) => void {
         return (req, res, next) => {
@@ -17,7 +17,7 @@ export class OwnerAuthManager {
                     ).withStatus(401)
                 );
             } else {
-                let userId = req.route.params[idParam] || req.param(idParam);
+                let userId = req.params[idParam] || req.param(idParam);
                 if (req.user_id != userId) {
                     HttpResponseSender.sendError(
                         req, res,
@@ -44,7 +44,7 @@ export class OwnerAuthManager {
                     ).withStatus(401)
                 );
             } else {
-                let userId = req.route.params[idParam] || req.param(idParam);
+                let userId = req.params[idParam] || req.param(idParam);
                 let roles = req.user != null ? req.user.roles : null;
                 let admin = _.includes(roles, 'admin');
                 if (req.user_id != userId && !admin) {