+++ src/auth/RoleAuthorizer.ts
@@ -4,7 +4,7 @@ const _ = require('lodash');
 import { UnauthorizedException } from 'pip-services3-commons-node';
 import { HttpResponseSender } from '../services/HttpResponseSender';
 
-export class RoleAuthManager {
+export class RoleAuthorizer {
 
     public userInRoles(roles: string[]): (req: any, res: any, next: () => void) => void {
         return (req, res, next) => {