+++ test/services/DummyRestService.ts
@@ -75,7 +75,7 @@ export class DummyRestService extends RestService {
     }    
         
     public register() {
-        this.registerMiddleware('/dummies', this.incrementNumberOfCalls);
+        this.registerInterceptor('/dummies', this.incrementNumberOfCalls);
 
         this.registerRoute(
             'get', '/dummies',